pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ваше Имя <email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"